chore: [분리] 뷰 분리, 폴더 정리
버튼은 button에, 뷰는 view에 파일 저장
ProfileView -> top, bottom으로 분리해서 구현
